+++ packages/components/src/components/telekom/app-header/app-header.tsx
@@ -48,6 +48,7 @@ export class Header {
   @Prop() logoHref?: string;
   @Prop() logoTitle?: string;
   @Prop() logoClick?: any;
+  @Prop() logoAriaDescribedBy?: string;
   @Prop() claimLang: string;
   @Prop() portalName?: string = '';
   @Prop() mainNavigation?: any = [];
@@ -435,13 +436,18 @@ export class Header {
                   {this.hasSlotLogo ? (
                     <slot name="logo"></slot>
                   ) : (
-                    <app-logo
-                      claim
-                      claimLang={this.claimLang}
+                    <scale-logo
+                      transparent
+                      language={this.claimLang}
                       href={this.logoHref}
                       logoTitle={this.logoTitle}
                       onClick={this.logoClick}
-                    ></app-logo>
+                      variant="white"
+                      scrollIntoViewOnFocus={true}
+                      focusable={true}
+                      styles=":host { --logo-size: 36px;} @media (max-width: 1023px) { :host {--logo-size: 26px;} }"
+                      logoAriaDescribedBy={this.logoAriaDescribedBy}
+                    ></scale-logo>
                   )}
                 </div>
                 <div class="header__brand-sector">{this.menuSector()}</div>
@@ -453,13 +459,16 @@ export class Header {
               <span class="header__nav-after"></span>
               <div class="header__nav-content">
                 <div class="header__nav-logo">
-                  <app-logo
-                    color="#e20074"
+                  <scale-logo
+                    transparent
+                    language=""
                     href={this.logoHref}
                     logoTitle={this.logoTitle}
                     onClick={this.logoClick}
-                    focusable={this.scrolled}
-                  ></app-logo>
+                    focusable={this.scrolled || this.sticky}
+                    size={24}
+                    logoAriaDescribedBy={this.logoAriaDescribedBy}
+                  ></scale-logo>
                 </div>
                 <div class="header__nav-menu-wrapper">
                   <div class="header__nav-menu-main">{this.menuMain()}</div>